snet: Remove sciond dependency completely
Contributes scionproto/scion#3348
