+++ go/lib/snet/BUILD.bazel
@@ -24,7 +24,6 @@ go_library(
         "//go/lib/hpkt:go_default_library",
         "//go/lib/l4:go_default_library",
         "//go/lib/log:go_default_library",
-        "//go/lib/sciond:go_default_library",
         "//go/lib/scmp:go_default_library",
         "//go/lib/serrors:go_default_library",
         "//go/lib/snet/internal/ctxmonitor:go_default_library",
