+++ go/lib/snet/path.go
@@ -15,15 +15,11 @@
 package snet
 
 import (
-	"crypto/sha256"
-	"encoding/binary"
 	"net"
 	"time"
 
 	"github.com/scionproto/scion/go/lib/addr"
 	"github.com/scionproto/scion/go/lib/common"
-	"github.com/scionproto/scion/go/lib/sciond"
-	"github.com/scionproto/scion/go/lib/serrors"
 	"github.com/scionproto/scion/go/lib/spath"
 )
 
@@ -68,111 +64,6 @@ type Path interface {
 type PathInterface interface {
 	ID() common.IFIDType
 	IA() addr.IA
-}
-
-var _ Path = (*path)(nil)
-
-type path struct {
-	// sciondPath contains SCIOND-related path metadata.
-	sciondPath *sciond.PathReplyEntry
-	// spath is the raw SCION forwarding path.
-	spath *spath.Path
-	// overlay is the intra-AS next-hop to use for this path.
-	overlay *net.UDPAddr
-	// source is the AS where the path starts.
-	source addr.IA
-}
-
-// NewPathFromSDReply creates a snet.Path from a sciond.PathReplyEntry. It
-// should only be used to refactor code depending on SCIOND into code depending
-// on snet.
-func NewPathFromSDReply(srcIA addr.IA, replyEntry *sciond.PathReplyEntry) (Path, error) {
-	if replyEntry == nil {
-		return &path{source: srcIA}, nil
-	}
-	sp := spath.New(replyEntry.Path.FwdPath)
-	// Preinitialize offsets, we don't want to propagate unusable paths
-	if err := sp.InitOffsets(); err != nil {
-		return nil, serrors.WrapStr("path error", err)
-	}
-	overlayAddr := replyEntry.HostInfo.Overlay()
-	return &path{
-		sciondPath: replyEntry,
-		spath:      sp,
-		overlay:    overlayAddr,
-		source:     srcIA,
-	}, nil
-}
-
-func (p *path) Fingerprint() string {
-	if p.sciondPath == nil {
-		return ""
-	}
-	h := sha256.New()
-	for _, iface := range p.sciondPath.Path.Interfaces {
-		binary.Write(h, common.Order, iface.RawIsdas)
-		binary.Write(h, common.Order, iface.IfID)
-	}
-	return string(h.Sum(nil))
-}
-
-func (p *path) OverlayNextHop() *net.UDPAddr {
-	return copyUDP(p.overlay)
-}
-
-func (p *path) Path() *spath.Path {
-	if p.spath == nil {
-		return nil
-	}
-	return p.spath.Copy()
-}
-
-func (p *path) Interfaces() []PathInterface {
-	if p.spath == nil {
-		return nil
-	}
-	res := make([]PathInterface, 0, len(p.sciondPath.Path.Interfaces))
-	for _, intf := range p.sciondPath.Path.Interfaces {
-		res = append(res, intf)
-	}
-	return res
-}
-
-func (p *path) Destination() addr.IA {
-	if p.sciondPath == nil {
-		return p.source
-	}
-	return p.sciondPath.Path.DstIA()
-}
-
-func (p *path) MTU() uint16 {
-	if p.sciondPath == nil {
-		return 0
-	}
-	return p.sciondPath.Path.Mtu
-}
-
-func (p *path) Expiry() time.Time {
-	if p.sciondPath == nil {
-		return time.Time{}
-	}
-	return p.sciondPath.Path.Expiry()
-}
-
-func (p *path) Copy() Path {
-	return &path{
-		sciondPath: p.sciondPath.Copy(),
-		spath:      p.spath.Copy(),
-		overlay:    copyUDP(p.overlay),
-		source:     p.source,
-	}
-}
-
-func (p *path) String() string {
-	if p.sciondPath == nil {
-		return ""
-	}
-	return p.sciondPath.String()
 }
 
 // partialPath is a path object with incomplete metadata. It is used as a
