+++ go/lib/snet/snetmigrate/BUILD.bazel
@@ -3,6 +3,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "go_default_library",
     srcs = [
+        "path.go",
         "querier.go",
         "resolver.go",
     ],
@@ -14,7 +15,9 @@ go_library(
         "//go/lib/pathmgr:go_default_library",
         "//go/lib/pathpol:go_default_library",
         "//go/lib/sciond:go_default_library",
+        "//go/lib/serrors:go_default_library",
         "//go/lib/snet:go_default_library",
+        "//go/lib/spath:go_default_library",
         "//go/lib/spath/spathmeta:go_default_library",
     ],
 )
