+++ go/tools/showpaths/BUILD.bazel
@@ -14,6 +14,7 @@ go_library(
         "//go/lib/sciond/pathprobe:go_default_library",
         "//go/lib/serrors:go_default_library",
         "//go/lib/snet:go_default_library",
+        "//go/lib/snet/snetmigrate:go_default_library",
     ],
 )
